chore: change dashboard page to home page
