--- src/pages/Dashboard/Dashboard.stories.tsx
@@ -1,9 +1,0 @@
-import { Meta, StoryObj } from "@storybook/react";
-import { Dashboard } from "./Dashboard";
-
-const meta: Meta<typeof Dashboard> = { component: Dashboard };
-export default meta;
-
-type Story = StoryObj<typeof Dashboard>;
-
-export const Default: Story = { render: Dashboard };
